Tests/LibWeb/Text/input/hit_testing/text.html

<style>
    * {
        border: 1px solid black;
    }
</style>
Lorem ipsum dolor sit amet, consectetur adipiscing elit. Quisque eu ante est. Integer ipsum sem,
tincidunt quis felis quis, efficitur fringilla neque. Aliquam erat volutpat. Donec feugiat euismod
sapien. Donec vel egestas arcu. Suspendisse luctus rhoncus mi quis elementum. Maecenas vel nisi
maximus, viverra tellus quis, ultrices elit. Quisque congue velit quis lectus congue, ut consectetur
nulla pharetra. Nullam euismod leo eget magna auctor, et bibendum urna tincidunt. Morbi molestie
gravida ex ac consectetur. Duis pretium gravida augue eu sagittis.
<script src="../include.js"></script>
<script>
    test(() => {
        const hitTestingResult = internals.hitTest(50, 50);
        println(hitTestingResult.indexInNode);
    });
</script>
